# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Peexoo's API swagger documentation",
        "description": "This is a sample Petstore server.  You can find out more about Swagger at [http://swagger.io](http://swagger.io) or on [irc.freenode.net, #swagger](http://swagger.io/irc/).",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "apiteam@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/subscription/charge/{subscription_plan_id}": {
            "get": {
                "tags": [
                    "subscription"
                ],
                "summary": "Generates Payment for an authenticated user to a Subscription Object",
                "operationId": "requestPaymentForSubscription",
                "parameters": [
                    {
                        "name": "subscription_plan_id",
                        "in": "path",
                        "description": "Id of the Subscription to join",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                },
                "security": [
                    {
                        "api_key_scheme": []
                    }
                ]
            }
        },
        "/api/subscription/complete": {
            "post": {
                "tags": [
                    "subscription"
                ],
                "summary": "This endpoint has to be hit by the payment gateway, but we'll make it open for testing purposes.",
                "operationId": "subscribeUser",
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                },
                "security": [
                    {
                        "api_key_scheme": []
                    }
                ]
            }
        },
        "/api/subscription/search": {
            "post": {
                "tags": [
                    "subscription"
                ],
                "summary": "Searches for an authenticated user's subscriptions",
                "operationId": "getUserSubscriptions",
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                },
                "security": [
                    {
                        "api_key_scheme": []
                    }
                ]
            }
        },
        "/api/subscription/package": {
            "patch": {
                "tags": [
                    "subscription"
                ],
                "summary": "Adds a new Subscription package",
                "operationId": "createSubscriptionPackage",
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                },
                "security": [
                    {
                        "api_key_scheme": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "api_key_scheme": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}